backend/requirements.txt
paramiko
python-dotenv
groq
web3
